pyproject.toml
# --- Build system (для pip install -e .) ---

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=68", "wheel"]

[dependency-groups]
dev = [
  "toml-sort (>=0.24.3,<0.25.0)",
  "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)"
]

[project]
authors = [
  {name = "hkutomanov"}
]
dependencies = [
  "matplotlib~=3.10",
  "ortools~=9.14",
  "pandas~=2.3",
  "pydantic~=2.5",
  "pyyaml~=6.0"
]
description = "Opmed optimization framework"
license = "MIT"
name = "opmed-optimization"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "black~=25.9",
  "hypothesis~=6.142",
  "mypy~=1.18",
  "pre-commit~=4.3",
  "pytest-cov~=7.0",
  "pytest~=8.0",
  "ruff~=0.14",
  "toml-sort~=0.24.3"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
exclude = "(data/|docs/|schemas/|infra/|scripts/)"
explicit_package_bases = true
ignore_missing_imports = true
mypy_path = ["src"] # src в sys.path
python_version = "3.11"
strict = true

# --- Poetry metadata (чтобы не потерять совместимость) ---
[tool.poetry]
packages = [{from = "src", include = "opmed"}]

[tool.pytest]

[tool.pytest.ini_options]
addopts = "-q --cov=opmed --cov-report=term-missing"
testpaths = ["tests"]

# --- Конфигурации инструментов ---
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
ignore = ["E501", "UP017"]
select = ["E", "F", "I", "UP"]

# --- src-layout для setuptools ---
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
